created  2 new class hierarchies, added 2 new abstract classes, implemented 2 new interfaces and aded exception handling together with 3 new manager classes and improving a fourth.
Well where do i start with this one, i haven't been able to sleep last night so i decide to code for a bit and now its 1:32 PM the next day. I have added a lot of functionality to the projects from adding the vehicle hierarchy that adds the car class and 5 new subclasses for it and  the user hierarchy  to adding a whole new suite of manager classes to handle everything necessary for the operation of all key classes. On top of that i have changed a lot of the file structure of the project in order to make it more easily readable and in order to improve reusability i made an abstract user class class and an abstract vehicle class as well as 2 new interfaces. i also decided to add a bit of exception handling in order to make the project as realistically useful as i can at the current stage.